Bug#22165 - crash while doing ALTER in parallel with another transactions Eliminated races.
Race condition 1.
ALTER/DROP/TRUNCATE have to ensure that table has no committed records
that are not yet "write complete". Otherwise crash is possible if table is 
deleted, but then accessed during Transaction::writeComplete() in Record::poke()

It is now solved, Storage::external_lock will wait for all transaction on 
the table  to become write complete, if there is no uncommited records.

Race condition 2.
In Transaction::commit(), Transaction state is set to Committed to early
while the transaction is still in the active list. This can produce incorrect
results in Transaction::hasUncommitedRecords, DROP is not blocked.
When later Transaction::commit() traverses the record list, it can crash 
accessing freed memory in Table::updateRecord().

This is fixed and transaction state is changed during the transition from 
active list to committed list.