Bug#35117 - Backup: Server crash for backup of CSV engine with CHAR data type Bug 33566 - Backup: crash with partitions and Falcon Bug 36792 - Assertion table->in_use == current_thd in Field_string::store, line 6259
Patch attached to 35117 but applies to all three bugs.
            
Before: With default driver, the table->in_use variable is set
to point to the locking thread when a table is locked. The
variable points to the wrong thread (i.e., the locking 
thread) when backup later tries to get the data from the table.
            
Now: Before getting data from a table, the default driver sets
the table->in_use pointer to the backup thread.