Bug#38843 Falcon does not start if datadir is located on home directory on Solaris
Problem
Falcon tries to create all directories recursively before creating a file
While trying to create a directory structure, IO::createPath runs onto Solaris bug. 
mkdir returns ENOSYS instead of EEXIST on an existing path, if if it happens to be NFS 
automounted root dir.

Fix: 
1) IO::createPath - Workaround the bug by handling ENOSYS like EEXIST. 
We do not expect any other OS to return ENOSYS(not supported) for mkdir().


2) As Falcon can always rely on datadir to exist, enclose all current 
occurences of  IO::createPath() into #ifndef FALCONDB. It is possibly 
needed for Netfrastructure, so the change is safe.