Bug #38933 - crash if rollback and concurrent DDL operations
Problem : Transaction record list is modified and entries of
the list are freed during Transaction::rollback.If there is 
a parallel execution of DDL statement, DDL  will traverse the 
list and can crash accessing freed memory

Fix: protect record list on Transaction::rollback. The syncObject
for this purpose has always been syncIndexes and it is used here
once again. syncIndexes usage is possibly a hack. TODO: have 
dedicated syncObject for protecting the list.