Bug#38933 - Falcon crash in Transaction::hasRecords during concurrent DDL
The problem is that Transaction::hasRecords traverses  transaction's record
list while a concurrent operation (rollback, writeComplete,addRecord) may 
concurrently modify it.
     
This patch removed whacko-reuse of the Transaction::syncIndexes to protect
the record list and introduces new lock Transaction::syncRecords for exactly 
this purpose.
      
Hopefully all places where the list is read or modified are now protected 
by this lock,except Transaction destructor, where there should be nobody 
accessing the list concurrently. When the list is freed, firstRecord  is 
set to NULL to prevent crash in parallel Transaction::hasRecords().