Bug #38947 UPDATE threads in endless Table::fetchForUpdate loop = livelock
Problem: in Table::fetchForUpdate() there is a small possibility for a 
race condition -  if record belongs to transaction that is being committed
currently and the state of this transaction is still Active, but syncActive
is already unlocked. This causes re-fetch() in the fetch thread without 
any wait, instead of waiting for falcon_lock_wait_timeout seconds.

This is fixed by moving signaling  waiters via syncActive.unlock(), 
after transaction state has changed from active to committed.