WL#4296 : Online Backup : Refine backup progress mechanism
PATCH 1 of 2 : Code changes.

This worklog changes the existing backup progress mechanism from MyISAM
tables to logging capability similar to the query log. The new mechanism
allows users to specify the destination of the log as a file, table, both,
and allows users to turn the logs on or off.