Bug#39153 Failing to lookup a host name can lead to crash in current IPv6 implementation
The reverse name lookup used to ensure the server that the client is who it says to be can
fail. When it does it crashes because of an additional freeaddrinfo on an uninitialized
structure.

When a host name is added to the host name cache we do a memcpy on sockaddr_storage
structures but uses the size of struct addrinfo. This also caused a crash.