Bug#39138 : Falcon total database deletion after CREATE + DROP TABLESPACE + recovery
Problem: Falcon misinterpretins some errors from openDatabase(), assumes database does 
not exist and tries to create a new database. During creation, it  overwrites already existing 
files (by passing O_CREAT|O_TRUNC to open() on Posix, and CREATE_ALWAYS to CreateFile on Windows).

This is corrected in this patch 

- open() is now called with O_CREAT|O_EXCL  and CreateFile() with CREATE_NEW. 
This will make Falcon never overwrite existing files.

- To ensure the proper cleanup of files created during create database attempt, 
"create" status is  tracked. On error, createDatabase sets global deleteFilesOnExit 
flag - this causes new files to be deleted in destructors of IO and SerialLogFile. 
But there is no cleanup if process terminates with exception/signal.
A more elegant solution for cleanup is possible on file systems that support user-mode 
transactions (like recent NTFS), by making the whole createDatabase/openDatabase a single 
transaction with commit() at the end.  Unfortunately this solution is not portable yet,
not even on downlevel  Windows ,so it is not implemented.

- We also ensure with ASSERT, that createDatabase never opens existing files, but always 
creates new ones.
- We do not attempt to create a database after failing recovery anymore.