Bug#39189 - Backup: backup_backupdir.test fails in pushbuild
Old behavior: When giving hard path to backup command, a '/' was 
appended at the end of the path (i.e., after the filename). 
This was because the whole path was converted to fit the system. 
In addition, convert_dirname crashed on some but not all 
platforms when length was not specified.
            
New behavior: Only the directory name is converted to system. The
filename is then appended without conversion. Conversion is 
done with explicit path length.