This is a fix for bug#38304 - Data contents becomes NULL if column name in Falcon table uses accented letters.
In StorageInterface::encodeRecord the value for fields with
international character(s) in their names are coded as
NULL (encodeNULL). This is because there is no entry for these
fields in the fieldMap. The values for these fields are therefore
never written to the database.
      
When populating the fieldMap ( in StorageInterface::mapFields),
the fieldId of every field is looked up by calling
StorageTableShare::getFieldId, which in turn calls
Table::findField.
      
Table::findField, is not able to match the field name looked up
in the SymbolManager (SymbolManager::getSymbol), with the field
name in the Table.
      
This is because SymbolManager::getSymbol tries to uppercase
every Symbol it looks up, but the uppercasing does not handle
international characters. The name it tries to look up, and
eventually returns, is the partial name of the field, up to the
first international character.
      
Because we "can't find" the correct field, a NULL entry is added
to the fieldMap for this field.
      
According to Jim Starkey, the fix for this would be to replace
the call to SymbolManager::getSymbol, with a call to
SymbolManager::getString, which does not uppercase the input.
      
This seems to work, and I have added a test-case that tests this
for different combinations of the SQL modes (also suggested by
Jim)
      
There are some twenty different SQL modes, which would result in
millions of different combinations, so I have picked the modes
that to the best of my knowledge, are the most interesting ones.