BUG#39127 Backup fails on PowerMac platform
The PowerMacG5 platform does not reference the variables used in the int8store() macro as 
originally written. The address of write_id is invalid forcing the system to crash.

This patch fixes the problem by using pointers to allocated memory for the int8store()
and uint8korr() as type uchar * instead of ulonglong.