BUG#38261 (Online backup: BACKUP command does implicit commit)
It was clarified that it is OK for BACKUP to perform implicit commit, so it 
is not a problem. Also, the server was modified so that SQL commands marked with 
CF_AUTO_COMMIT_TRANS flag automatically perform implicit commits at the 
beginning and at the end of execution. The BACKUP and RESTORE commands are 
already marked with the flag.

Hence, there is no need for doing any commits inside backup kernel code - calls 
to ha_autocommit_or_rollback() and end_active_trans() are removed from kernel.cc 
by this patch.

The patch adds new test backup_commit_backup which tests that BACKUP command 
performs implicit commit for the current transaction and does not interfere with 
transactions in other connections.