Bug#32858: Error: "Incorrect usage of UNION and INTO" does not take subselects into account
When parsing a new UNION branch, there is a check if the previous
branch contained INTO, which would be illegal syntax. But this is
only the case if the new branch is on the top-level.
Fixed in 6.0 by introducing a boolean argument so that the check
is disabled when parsing nested UNION expressions.