(cherry-pick from 6.0-bugteam repo) Bug 35987 - crash report on windows doesn't resolve stack traces. The problem here is that symbols can not be loaded, because symbol path is not set and  default path does not include the directory where PDB is located.
The problem is _not_ reproducible on the same machine where
mysqld.exe is built - if PDB is not found in the symbol path,
dbghelp would fallback to fully qualified PDB path as given in the
executable header and on the build host this will succeed.
      
The solution is to calculate symbol path and pass it to SymInitialize()
call.