Bug#37226 -  Explicit call of my_thread_init() on Windows for every new thread. Bug#33031 - app linked to libmysql.lib crash if run as service in vista under localsystem.
This patch completely removes DllMain() from the libmysql.dll in 6.0
eliminating explicit my_thread_init()/WSAStartup() during dll load
or thread attach. 

This patch is slightly different from what is done in 5.0/5.1.
The difference is that there is no way that to reactivate DllMain() 
code (in 5.x it is possible  with LIBMYSQL_DLLINIT environment variable)