A fix and a test case for Bug#36473 Assertion "! is_set()" fails in Diagnostics_area::set_ok_status with Falcon.
When running in strict mode, update_auto_increment() function
may return an error. In case of Bug#36473 the error was due to an 
out of range auto increment value for a TINYINT column.
Falcon handler implementation did not check for the return value
and that led to an attempt to commit statement transaction in the
server and return success status to the user.
Fix by checking the return value of update_auto_increment().