WL#2110 (Stored Procedures: Implement SIGNAL) WL#2265 (Stored Procedures: Implement RESIGNAL) Bug#11661 (Raising Exceptions from within stored procedures: Support for            SIGNAL statement)
Fixed review comments related to memory management.

- Changed all the handle_condition() methods to return a SQL_condition
as *output*, instead of *input*.
The logic now find first where the condition should be written to,
and then the caller writes the extra attributes in place.
This change removes the need for SQL_condition::deep_copy().
Also, when a condition is dropped (Internal handlers, warn_list full),
this saves some CPU cycle.

- Removed SQL_condition::deep_copy()

- Removed sp_rcontext::m_cond_root, and reverted related changes.

- Adjusted the code that raises a condition according to the new paradigm.