Bug#39916: pool-of-threads fails to increment the number of created threads
The problem is that threads created under the pool-of-threads mode
weren't being properly accounted for in the Threads_created variable.
The solution is to increase the number of Threads_created every time
a thread is created and added to the pool. This means that the number
of threads created is equal to the size of the thread pool.