Bug#39212 Falcon exception on recovery when one of the log files is empty
Reason is a bug in reiserfs file system.
The serial log file is initially 512 bytes.  On reiserfs,
for files opened with O_DIRECT, read() will fail if file size
is between 1 to 4095 returns error (EINVAL). The bug supposedly 
occurs because file size is smaller than page size.
            
Solved with a workaround: initial size of serial log files is 
now 8K (we assume largest page size we need to deal with is 
8KB).