This is a fix for bug#39342 (YEAR '=' comparison fails when index is present)
According to the reference documentation, YEARSs can be declared as
YEAR(2) or YEAR(4) to specify a display width of two or four 
characters. It turns out that YEARs are stored differently in Falcon 
depending on this. In the case of a YEAR(4), when searching
an index, the search keys given where off by -1900 years, and the 
search did therefore not find the expected result. 
This patch ensures that both types of YEARs are stored equally 
inside Falcon, in a format that matches the search keys given by 
the server. The change is limited to StorageInterface::encodeRecord
and StorageInterface::decodeRecord.

A regression test is also added as a part of this commit.