This commit introduces a new tests for the YEAR type in Falcon and is a part of the patch for bug#39342.
The test tests the 4 digit version of the YEAR type. This test should
pass with the fix provided in another commit for this patch.  The test
tests 2 digit, 4 digit, and character input in combination with indexes
and LIMITs. Also some functions are tested.

I have also created a similar test for the 2 digit version of YEAR. This test
will currently break due to a bug in the server (40171), and is
therefore not part of this commit, but will be attached to the bug
report (40171).