Fix for Bug#40135.
The test case was showing long run time on slower machines. I tested it on a idle 450MHz.
PPC machine and had a run time of ~ 5 minutes. It's perfectly possible to have a runtime >
15 minutes on busy machines.

Fixed by lowering the loop count to 100k. Also created a corresponding -big.test with loop
count of 1 million.