BUG#39714 mtr.check_warnings can be painfully slow if error log file is large
Test backup_logs.test mistakenly turned on debug output to error log
by setting debug="d", which caused the error log to be very large (> 1M)
and mtr.check_warnings would spend a long time to process the error log,
and can cause the test fail because of timeout. The test case does not
need to enable debug output, what it need is to reset the debug action
it set before.

This patch fixed the problem by resetting debug to "" instead of "d".