BUG#36664 Server crash in backup_no_be test on Solaris.
Enabling this test to see if this is still a problem.  
(Not able to reproduce it myself).

Also, cleaning up some debug settings in test to avoid generating large volumes of load.

Note that this test was disabled due to valgrind issues, but I could not
find that these issues had ever occurred in Pushbuild.