Bug #38186: CREATE TABLESPACE failed when invoked immediately following a  DROP TABLESPACE statement that used the same tablespace name.
 Fixed a long-standing problem that tablespace file was present even after "drop tablespace" successfully 
completed.  The problem was worked around previously, this patch contains refined solution. 
Previously, file was closed and deleted when the gopher thread processed dropTableSpace record.
 Now  the file is removed earlier, in the client thread . Gopher thread only closes the file descriptor later. 

This technique ensures that 
- datafile does not exist after "drop tablespace" is completed
- file descriptor remains valid as long as it can be used by commited ,  write-incomplete transactions.

On  Windows, removing an open file requires FILE_SHARE_DELETE flag for CreateFile. Also file has to be renamed to 
a unique name before deletion. This functionality is implemented  in posix-alike winOpen() and winUnlink()  functions

This patch also contains missing test case for Bug#38186.