Bug#40112 - DeferredIndexWalker::next() was not prepared for the first leaf bucket to be empty.  But it can become empty after several verbs in the same transaction change the same records.
When the same record is changed more than once in the same transaction, 
the older unneeded versions are deleted along with the index entries.  
Even though DeferredIndex key nodes can be deleted, the buckets are not.  

If the first leaf bucket is empty, DeferredIndexWalker::next() would 
return NULL on the first call.  In SRLUpdateIndex::append(), this would
result that nothing would be put into the Serial log at commit time for 
this DeferredIndex.  Then this DefferedIndex would not get put into 
index pages.

The solution is to only return NULL on the first call to 
DeferredIndexWalker::next() if the first leaf bucket is empty AND there 
is no parent bucket.  Otherwise, use the normal mechanism to advance to 
the next slot.