Bug #40302 error on tablespace creation is not handled cleanly
Problem: If tablespace.create() fails, system transaction that updates tablespace
info is not rolled back. It can be possibly commited with the next system transaction,
that will add a tablespace to system tables even if datafile does not exist.
      
Additionally, createTableSpace log record (indicating creation of datafile)  that logically 
belongs to the same transaction as  system tables and has the same transaction
id was wrongly written after "commit".
      
With this patch, 
 -  there is a  rollback in case of error and appropriate cleanup in the filesystem
 -  createTableSpace now belongs to the same transaction that changes system tables.