Bug#39696, "Assertion in Table.cpp (dup->state == recDeleted) fails during falcon_chill_thaw"
The record chill operation sometimes generates SRLUpdateRecord entries in the serial
log that have zero data length. This causes a corruption of the serial log.

The total data length of the chilled records is known only after all of the records have
been processed in SRLUpdateRecords::append(). If none of the record versions within
a given savepoint are eligible to be chilled, the total data length of the serial log
record will be zero.

Even when there are zero bytes of data to commit to disk, SRLUpdateRecords::append() issues
a serial log flush and forces the creation of a new serial log window. This creates a one-byte
offset error in the serial log that ultimately results in a fatal assertion during gopher
thread processing.

To prevent this error, SRLUpdateRecords::append() now only flushes serial log records with a
non-zero data length.