BUG#36804 - Falcon does not detect deleted tablespace files
Falcon silently recreates it's system data files if it wasn't
able to open them. With this fix Falcon provides some
additional information in error log.

When starting mysql server with Falcon and we were unable
to open Falcon system data files, write a warning to error
log.

Write an error message to error log when attempting to access
a table that doesn't exist in Falcon.