Bug#39789 : Falcon recovery failure after several CREATE + DROP TABLESPACE
The problem in this specific situation was removal of datafiles in SRLDropTable
space::redo(),even if tablespace with the same datafile path in use by another 
tablespace instance.
      
The fix reimplements create/dropTablespace handling in recovery in 
a manner consistent with the rest of Falcon recovery.
      
-  pass1(check) of recovery  collects information about what tablespaces 
   are dropped at the end

-  pass2 and redo skip any reference to dropped tablespace 
  (except for SRLDropTableSpace record)

-  tablespace (not marked as dropped) are created during 
   pass2 of recovery

-  tablspaces marked as dropped are "redropped" in phase2 of recovery 
   i.e removed from memory structures (hashtables and lists).
   Datafile that is possibly left over in previous session is deleted
   as well.

This patch moves tableSpaceId from classes derived from SerialLogRecord
to the parent class, for technical reasons - it is much simpler to check
if log record should be skipped in different recovery phases.