BUG#40262 (Binary log information incorrect in backup of no data with binlog            turned on)
This is a preliminary patch which fixes a problem in si_log.h which leads to a 
carash reported in the bug. According to specifications, methods 
Backup_log::binlog_file() and Backup_log::master_binlog_file() should do nothing 
if NULL string is passed as the argument. But they crashed instead (due to a 
call to strlen() on NULL string). This patch fixes this.