Fix to bug#39260 Falcon should detect that it has been compiled without support for C++ exception
This is the second patch for handling the problem of Falcon being compiled without support for
C++ exceptions. Changes how Falcon behaves when it detects that it has been compiled without 
exception support from returning an error code to the server (as implemeneted in first patch)
to crash the process. The rationale for this is that this should only happen in rare occations 
where configure and/or compile has been done with wrong compiler options.