Do not use Record::state = recUnlock.  It is not recognized by many parts of the code and can cause the wrong behavior.
Also, A lock record that is unlocked should stay 
on the transaction.  It is still replaced on the 
recordLeaf, but other threads with pointers to it 
from a concurrent Table::fetch need to be able to 
see the transaction.  It will be taken off the
transaction later with all the other records on 
the transaction.