Fix for Bug#35503 - falcon_bug_34351_C.test times out on Pushbuild. Reducing loop count inside procedure from 10k to 5k in both falcon_bug_34351_A and falcon_bug_34251_C, as per Kevin Lewis' recommendations. Test case running time is effectively cut in half by doing this.
Also adding new -big versions of those tests, running 20k iterations, in case 
the bug is not triggered by 5k iterations in the non-big version. 
These tests need to be run using the --big-test option of mysql-test-run.pl.
On slower machines one might have to increase the test case timeout as well,
e.g. --testcase-timeout=30 for a timeout of 30 minutes.