Bug#39076 - The test backup_nodata_driver fails with blackhole         and federated in build.
It was reported that the test fails on RESTORE with "Table
'bup_data.myisam1' doesn't exist". But this error was correct.
I added the expected error directive.

However, at first the test case did not start on Linux at all.
The example storage engine is built as a dynamically loadable
plugin only. I added the required prerequisites. But now the
test case is skipped on Windows. Currently we do not have
Windows support for dynamically loadable plugins.
See WL#3653 - Loadable Engine on Windows