Bug#39419 Falcon should not be compiled if compiler or options do not support exceptions
Adds a check to configure that verifies that the compiler with the
given compiler options will produce code that handles C++ exceptions. This
test is done by running a small program that throws an catches an exception.
If this program does not work correctly, configure will exit will an error 
message and needs to be re-run either with correct compiler options of with
Falcon disabled.
      
Adds a check to the Falcon source code that tests that GCC and Windows compilers are
compiling with C++ exceptions enabled. This test will fail during compilation if the
compiler is not running with C++ exceptions enabled.