BUG#40303 (Backup: Wrong error send to client if several errors reported) BUG#40304 (Backup: Errors reported during BACKUP show as warnings on the            error stack)
Before: Errors reported using backup::Logger class could be saved and in 
case several errors are reported, the last one was sent to client as an
error reply from BACKUP/RESTORE command.

After: Errors reported with backup::Logger are reported to the server 
using my_printf_error(..) function. This function pushes error on the 
error stack (unles told otherwise) and stores the first reported error
so that it is send to the client.

Additionally.

- Function log_error(..) is moved to the Logger class.

- Code is changed to consistently use Logger::report/log_error() for error 
reporting.

- Semantics of Backup_restore_ctx::fatal_error() has been changed to 
discourage its use as an error reporting facility. Now it is an internal helper 
method which moves the context object into error state. It is used only by the 
class methods and it is made private. The idea is that context object should 
move into error state only as a result of executing one of its methods (not 
externally).

- Improved error reporting in few places.