Bug #36897        Falcon: "use tablespace" is possible
The problem is that the code only checks for the
existence of the target filename when entering a
database. Added a check to verify that the target is
indeed a directory, and not a tablespace file.

Since the my_access() call only checks for file
existence, it is removed, since this is already
ensured with the my_stat() call.