Bug #41173  rpl_packet fails sporadically on pushbuild: query 'DROP TABLE t1' failed
This one is not a fix but a temporary addon to the test to print extra debug.
The test is also got narrowed to run in the row binlog format for time the bug is being fixed.

Todo: remove the debug prints and the format restriction within the real bug fix.