WL#4654 New transaction dependency manager for Falcon
Implements a new dependency manager for Falcon that uses start events and commit events for transactions
for determining the relative state between transactions and for determining when a transaction can
be removed. The transaction id is used as start event and a new "commitId" is used as commit event. 
Both the transaction id and the new commit id is generated from the transaction id sequence.
      
The old dependency manager is removed. This eliminates the states array and the dependency counter
from each transaction.