Bug#38426 - Backup of partitioned Falcon table will not use snapshot driver
Before, the Default backup driver was used when backing up a partitioned table. However, the Snapshot driver is capable of backing up partitioned tables if the underlying storage engine is InnoDB or Falcon.
      
With this patch, the Snapshot backup driver is used if the underlying storage engine of a partitioned table is InnoDB or Falcon