BUG#34754 Backup: Invalid error message for incorrect path name while performing backup
Will now use error messages generated by my_create (BACKUP) and my_open (RESTORE)
when failing to create/open a file since these message are more specific as to
the cause of the failure.

Note that the backup-specific errors will still be pushed on the error stack,
and it is still these errors that are recorded in the backup history log and 
in the error file.