WL#4654 New transaction dependency manager for Falcon
In the initial implementation of the new dependency manager all calls to 
Transaction::commitRecords() and purging of old transaction objects was done by the scavenger.
This patch extends this task to be done also on every commit (in future: only commit of 
non-read-only transactions) by calling a new version of TransactionManager::purgeTransactions() 
that does not set the locks on the active and committed transaction list but relies on that 
the caller has set these locks.

The reason for adding the call to this code in Transaction::commit() is that this is one
of the few places in the code where we have a lock on both the active and the commited transaction
lists.