WL#4654 New transaction dependency manager for Falcon
Patch 2 for optimization of read-only transactions: Optimize transaction
object allocations: introduce code
that checks if there is a transaction objects that is in state 
available in the active transaction list when starting a new transaction. 
It this is the case we alloacte this transaction object using a CAS
operation while only having a shared lock on the list. In order to have
available un-used transaction objects we pre-allocate some extra
(10) non-used transaction objects each time we need to set an exclusive
lock on the active transaction list.

The performance improvement is achieved by being able to allocate
a transaction object using only a shared lock on the active transaction
list.

A consequence of this change is that the active transaction list no longer
will can be assumed to be sorted on transaction id. As a consequence,
TransactionManager::purgeTransaction now needs to go through the list
to find the oldest transaction in the active list instead of just
checking the first transaction in the list.